+++ cmd/wuffs-c/internal/cgen/expr.go
@@ -333,6 +333,12 @@ func (g *gen) writeExprOther(b *buffer, n *a.Expr, rp replacementPolicy, pp pare
 			b.printf(")")
 			return nil
 		}
+		if isThatMethod(g.tm, n, t.IDReset, 0) {
+			// TODO: don't hard-code f_lzw.
+			b.writes("wuffs_gif__lzw_decoder__initialize(" +
+				"&self->private_impl.f_lzw, WUFFS_VERSION, 0)")
+			return nil
+		}
 		// TODO.
 
 	case t.IDOpenBracket: